requirements.txt
# Core ML dependencies (CPU-only versions for Lambda)
torch==2.6.0
torchvision==0.21.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
insightface==0.7.3
onnxruntime==1.16.3
onnx==1.15.0

# Image processing
pillow==10.0.1

# FastAPI and Lambda adapter
fastapi==0.104.1
mangum==0.17.0
python-dotenv==1.1.0

# AWS dependencies (boto3 is already in Lambda runtime)
elasticsearch==8.10.0

# Other dependencies
tqdm==4.66.1 